+++ packages/storage/src/providers/s3/utils/md5.native.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Md5 } from '@aws-sdk/md5-js';
-import { toBase64 } from '../AwsClients/S3/utils';
+import { toBase64 } from '../../../AwsClients/S3/utils';
 
 export const calculateContentMd5 = async (
 	content: Blob | string | ArrayBuffer | ArrayBufferView
