+++ packages/storage/src/providers/s3/utils/md5.ts
@@ -2,14 +2,18 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Md5 } from '@aws-sdk/md5-js';
-import { toBase64, utf8Encode } from '../AwsClients/S3/utils';
+import { toBase64, utf8Encode } from '../../../AwsClients/S3/utils';
 
 export const calculateContentMd5 = async (
-	content: Blob | string
+	content: Blob | string | ArrayBuffer | ArrayBufferView
 ): Promise<string> => {
 	const hasher = new Md5();
 	if (typeof content === 'string') {
 		hasher.update(content);
+	} else if (ArrayBuffer.isView(content) || content instanceof ArrayBuffer) {
+		const blob = new Blob([content]);
+		const buffer = await readFileToBase64(blob);
+		hasher.update(buffer);
 	} else {
 		const buffer = await readFileToBase64(content);
 		hasher.update(utf8Encode(buffer));
